scheduler/requirements.txt
# scheduler/requirements.txt
asyncio>=3.4.3
nats-py==2.5.0
httpx==0.25.0
asyncpg==0.29.0
python-dotenv==1.0.0
